Downloads/prism-agent-package/requirements.txt
semantic-kernel==1.29.0
torch
openai
python-dotenv
requests
easyocr
pillow
